- renaming to AccessibilityTagHelper and addressing PR comments
